requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
psycopg2-binary==2.8.6
bcrypt==3.1.7